requirements.txt
altgraph==0.17.4
certifi==2024.6.2
charset-normalizer==3.3.2
idna==3.7
packaging==24.1
pefile==2023.2.7
pillow==10.4.0
pyinstaller==6.8.0
pyinstaller-hooks-contrib==2024.7
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
pystray==0.19.5
pywin32-ctypes==0.2.2
requests==2.32.3
setuptools==70.2.0
six==1.16.0
urllib3==2.2.2
